fix(Expense): refactor process buttons
